refactor: providers.tsx의 인증 로직을 FSD에 맞게 분리
