+++ app/layout.tsx
@@ -1,6 +1,7 @@
 import './globals.css'
 
 import { Providers } from '@/shared/providers/providers'
+import { SessionSyncProvider } from '@/features/auth/session/ui/session-sync-provider'
 import AppToaster from '@/shared/ui/app-toaster'
 import { Geist, Geist_Mono } from 'next/font/google'
 import type { Metadata } from 'next'
@@ -31,8 +32,10 @@ export default function RootLayout({
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <Providers>
-          <AppToaster />
-          {children}
+          <SessionSyncProvider>
+            <AppToaster />
+            {children}
+          </SessionSyncProvider>
         </Providers>
       </body>
     </html>
