+++ src/features/auth/index.ts
@@ -4,22 +4,22 @@
 export { useLoginActions } from './login/model/use-login-actions'
 export { useLoginForm } from './login/model/use-login-form'
 export { default as LoginForm } from './login/ui/login-form'
-
 // Types (Auth feature specific types only)
 export type {
   LoginData,
   LoginFormData,
   SignupFormData,
 } from './model/types'
-
 // Password Reset
 export type { PasswordResetData } from './model/validation-schema'
 export { usePasswordResetActions } from './password-reset/model/use-password-reset-actions'
 export { default as PasswordResetDialog } from './password-reset/ui/password-reset-dialog'
-
 // Profile Avatar
 export type { ProfileAvatarProps } from './profile-avatar'
 export { ProfileAvatar, useProfileAvatar } from './profile-avatar'
+// Session
+export { useSessionSync } from './session/model/use-session-sync'
+export { SessionSyncProvider } from './session/ui/session-sync-provider'
 
 // Social Login Features (분리됨)
 export * from './social-github'
